+++ src/main/java/com/hzl/hadoop/workflow/vo/ApproveHistoryGatewayVO.java
@@ -42,7 +42,7 @@ public class ApproveHistoryGatewayVO implements Serializable {
 	/**
 	 * 审批动作，1同意，2拒绝，3跳过，4转交
 	 */
-	private Long approveAction;
+	private Integer approveAction;
 	/**
 	 * 租户id
 	 */
